Add files via upload
After labeling the images, my model was created with the name best.pt. Then, the model was run inside Visual Studio Code using Python. Additionally, I developed a control panel in Python to receive outputs from the model and control the robotic arm (ROARM-M1). The control panel was connected to both the model and the robot. Now, the robot is intelligent, capable of detecting the target object and moving towards it by receiving its 3D position.